+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTest.php
@@ -39,6 +39,7 @@ use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpClient\ScopingHttpClient;
 use Symfony\Component\HttpKernel\DependencyInjection\LoggerPass;
+use Symfony\Component\Mailer\Messenger\Middleware\SendMailAfterCurrentBusMiddleware;
 use Symfony\Component\Messenger\Transport\TransportFactory;
 use Symfony\Component\PropertyAccess\PropertyAccessor;
 use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
@@ -638,25 +639,41 @@ abstract class FrameworkExtensionTest extends TestCase
 
         $this->assertTrue($container->has('messenger.bus.commands'));
         $this->assertSame([], $container->getDefinition('messenger.bus.commands')->getArgument(0));
-        $this->assertEquals([
+
+        $expectedMiddleware = [
             ['id' => 'add_bus_name_stamp_middleware', 'arguments' => ['messenger.bus.commands']],
             ['id' => 'reject_redelivered_message_middleware'],
+            ['id' => 'send_mail_after_current_bus'],
             ['id' => 'dispatch_after_current_bus'],
             ['id' => 'failed_message_processing_middleware'],
             ['id' => 'send_message'],
             ['id' => 'handle_message'],
-        ], $container->getParameter('messenger.bus.commands.middleware'));
+        ];
+
+        if (!class_exists(SendMailAfterCurrentBusMiddleware::class)) {
+            array_splice($expectedMiddleware, 2, 1);
+        }
+        $this->assertEquals($expectedMiddleware, $container->getParameter('messenger.bus.commands.middleware'));
+
         $this->assertTrue($container->has('messenger.bus.events'));
         $this->assertSame([], $container->getDefinition('messenger.bus.events')->getArgument(0));
-        $this->assertEquals([
+
+        $expectedMiddleware = [
             ['id' => 'add_bus_name_stamp_middleware', 'arguments' => ['messenger.bus.events']],
             ['id' => 'reject_redelivered_message_middleware'],
+            ['id' => 'send_mail_after_current_bus'],
             ['id' => 'dispatch_after_current_bus'],
             ['id' => 'failed_message_processing_middleware'],
             ['id' => 'with_factory', 'arguments' => ['foo', true, ['bar' => 'baz']]],
             ['id' => 'send_message'],
             ['id' => 'handle_message'],
-        ], $container->getParameter('messenger.bus.events.middleware'));
+        ];
+
+        if (!class_exists(SendMailAfterCurrentBusMiddleware::class)) {
+            array_splice($expectedMiddleware, 2, 1);
+        }
+        $this->assertEquals($expectedMiddleware, $container->getParameter('messenger.bus.events.middleware'));
+
         $this->assertTrue($container->has('messenger.bus.queries'));
         $this->assertSame([], $container->getDefinition('messenger.bus.queries')->getArgument(0));
         $this->assertEquals([